ai-content-agent/backend/requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
psycopg2==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
httpx==0.25.2
